requirements.txt
streamlit
openai>=1.6.0
requests
pillow
